Rework vertx-rx tests to reduce flakiness
